Добавил изменение масштаба сцены на графической панели по колёсику мыши. Костыльно, но всё же. Сделал точки-координаты Point2D.Double.

+++ src/com/codebind/Shapes/Edge.java
@@ -36,6 +36,6 @@ public class Edge implements Drawable {
     @Override
     public void draw(Graphics2D g) {
         g.setColor(Color.BLUE);
-        g.drawLine(sourceNode.getPosition().x, sourceNode.getPosition().y, destNode.getPosition().x, destNode.getPosition().y);
+        g.drawLine((int)(sourceNode.getPosition().x), (int)(sourceNode.getPosition().y), (int)(destNode.getPosition().x), (int)(destNode.getPosition().y));
     }
 }
